+++ src/types.ts
@@ -1,5 +1,5 @@
 import type { OpenAPIV3 } from 'openapi-types'
-import  {SwaggerUIOptions} from 'swagger-ui'
+import type { SwaggerUIOptions } from 'swagger-ui'
 
 export interface ElysiaSwaggerConfig<Path extends string = '/swagger'> {
     /**
@@ -14,9 +14,9 @@ export interface ElysiaSwaggerConfig<Path extends string = '/swagger'> {
     >
     /**
      * Version to use for swagger cdn bundle
-     * 
+     *
      * @see unpkg.com/swagger-ui-dist
-     * 
+     *
      * @default 4.18.2
      */
     version?: string
@@ -41,12 +41,34 @@ export interface ElysiaSwaggerConfig<Path extends string = '/swagger'> {
     /**
      * Options to send to SwaggerUIBundle
      * Currently, options that are defined as functions such as requestInterceptor
-     * and onComplete are not supported. 
+     * and onComplete are not supported.
      */
-    swaggerOptions?: Omit<Partial<SwaggerUIOptions>, 
-                     'dom_id'|'dom_node'|'spec'|'url'|'urls' 
-                     |'layout' | 'pluginsOptions' | 'plugins'|'presets'
-                     |'onComplete' |'requestInterceptor'|'responseInterceptor'
-                     |'modelPropertyMacro'|'parameterMacro'
-                     >   
+    swaggerOptions?: Omit<
+        Partial<SwaggerUIOptions>,
+        | 'dom_id'
+        | 'dom_node'
+        | 'spec'
+        | 'url'
+        | 'urls'
+        | 'layout'
+        | 'pluginsOptions'
+        | 'plugins'
+        | 'presets'
+        | 'onComplete'
+        | 'requestInterceptor'
+        | 'responseInterceptor'
+        | 'modelPropertyMacro'
+        | 'parameterMacro'
+    >
+    /**
+     * Custom Swagger CSS
+     */
+    theme?: string | {
+        light: string
+        dark: string
+    }
+    /**
+     * Using poor man dark mode 😭
+     */
+    autoDarkMode?: boolean
 }